move map_addresses out
