+++ src/genjax/_src/core/generative/choice_map.py
@@ -935,34 +935,3 @@ def choice_map_filtered(
     c: ChoiceMap,
 ) -> ChoiceMap:
     return choice_map_empty if c.static_is_empty() else FilteredChm(selection, c)
-
-
-@Pytree.dataclass
-class AddrMapChm(ChoiceMap):
-    c: ChoiceMap = Pytree.field()
-    mapping: dict = Pytree.static()
-
-    @typecheck
-    @classmethod
-    def create(cls, c: ChoiceMap, mapping: dict) -> ChoiceMap:
-        if c.static_is_empty():
-            return choice_map_empty
-        return cls(c, mapping)
-
-    def get_value(self) -> Bool | BoolArray:
-        mapped = self.mapping.get((), ())
-        if mapped:
-            submap = self.c.get_submap(mapped)
-            return submap.get_value()
-        else:
-            return self.c.get_value()
-
-    def get_submap(self, addr: AddressComponent) -> ChoiceMap:
-        if ... in self.mapping:
-            mapped = self.mapping[...]
-            return self.c.get_submap(mapped).get_submap(addr)
-        else:
-            mapped = self.mapping.get(addr, addr)
-            if mapped is ...:
-                return self.c
-            return self.c.get_submap(mapped)
